requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.14
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.18
langchain-anthropic==1.4.3
langchain-core==1.3.3
langchain-groq==1.1.2
langchain-protocol==0.0.15
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.3
mcp==1.27.1
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.28
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.4.2
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uuid_utils==0.14.1
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.7.0
zstandard==0.25.0
